+++ SONIC_JK/jk_Act1_BG_object.cpp
@@ -42,7 +42,8 @@ namespace jk
 			Vector2 pos = tr->GetPos();
 			Vector2 mpos = pos - (Camera::CaluatePos(pos) - pos);
 		
-			TransparentBlt(hdc, 0, 0, 1200, 840, mImage->GetHdc(), mpos.x, mpos.y, 1200, 840, RGB(13, 72, 7));
+			TransparentBlt(hdc, 0, 0, 1200, 840, mImage->GetHdc(), 
+				static_cast<int>(mpos.x), static_cast<int>(mpos.y), 1200, 840, RGB(13, 72, 7));
 	}
