+++ SONIC_JK/jk_Ending.cpp
@@ -33,7 +33,7 @@ namespace jk
 	{
 		Transform* tr = GetComponent<Transform>();
 		Vector2 pos = tr->GetPos();
-		BitBlt(hdc, pos.x, pos.y, mImage->GetWidth(), mImage->GetHeight(), mImage->GetHdc(), 0, 0, SRCCOPY);
+		BitBlt(hdc, static_cast<int>(pos.x), static_cast<int>(pos.y), mImage->GetWidth(), mImage->GetHeight(), mImage->GetHdc(), 0, 0, SRCCOPY);
 
 		Gameobject::Render(hdc);
 	}
