+++ SONIC_JK/jk_Jeep_line_Handle.cpp
@@ -30,7 +30,7 @@ namespace jk
 	{
 		mImage = Resources::Load<Image>(L"Jeep_line_Head", L"..\\Resources\\Rock_Platform.bmp");
 		mAnimator = AddComponent<Animator>();
-		mAnimator->CreateAnimation(L"Jeep_line_Head", mImage, Vector2(73, 562), Vector2(24, 8), Vector2(0, 0), 1, 1, 1, Vector2::Zero, 0.1);
+		mAnimator->CreateAnimation(L"Jeep_line_Head", mImage, Vector2(73, 562), Vector2(24, 8), Vector2(0, 0), 1, 1, 1, Vector2::Zero, 0.1f);
 
 		mAnimator->Play(L"Jeep_line_Head", false);
 
@@ -53,7 +53,7 @@ namespace jk
 
 		if (mDir_x == 1)
 		{
-			angle += 10 * Time::DeltaTime() * 10;
+			angle += 10 * static_cast<float>(Time::DeltaTime())* 10.f;
 			if (angle > 135)
 			{
 				mDir_x = -1;
@@ -61,14 +61,14 @@ namespace jk
 		}
 		else if (mDir_x == -1)
 		{
-			angle -= 10 * Time::DeltaTime() * 10;
+			angle -= 10 * static_cast<float>(Time::DeltaTime()) * 10.f;
 			if (angle < 45)
 			{
 				mDir_x = 1;
 			}
 		}
 
-		float radian = angle * 3.14f / 180.0;
+		float radian = angle * 3.14f / 180.0f;
 
 		float x = centerX + radius * cos(radian);
 		float y = centerY + radius * sin(radian);
